requirements.txt
beautifulsoup4==4.11.2
dataclasses_json==0.5.7
dataclasses_json==0.5.8
load_dotenv==0.1.0
microdata==0.8.0
pymongo==4.3.3
python-dotenv==1.0.0
redis==4.5.5
Requests==2.31.0
